refactor:updated the UI of the welcome page
