+++ src/App.tsx
@@ -1,8 +1,8 @@
 import { BrowserRouter, Route, Routes } from 'react-router-dom';
 
+import SignIn from './pages/Welcome';
+
 import './App.css';
-import SignIn from './pages/signIn';
-import { routes } from './routes';
 
 const App = () => {
   return (
