--- src/components/GetLogIn/index.tsx
@@ -1,9 +1,0 @@
-import './index.css';
-
-export const GetLogIn = () => {
-  return (
-    <div className="getLogIn-attributes">
-      <button className="signIn-button">Sign In</button>
-    </div>
-  );
-};
