--- src/components/GetSignIn/index.css
@@ -1,17 +1,0 @@
-.email-value {
-  background: black;
-  padding: 10px 30px;
-  color: #fff;
-  z-index: 1;
-  font-size: 18px;
-}
-
-.getstarted-button {
-  background-color: red;
-  border: none;
-  color: #fff;
-  margin-left: 25px;
-  padding: 10px 20px;
-  font-size: 18px;
-  cursor: pointer;
-}
