+++ src/pages/index.ts
@@ -1,1 +1,1 @@
-export * from './signIn';
+export * from './Welcome';
