+++ src/routes.ts
@@ -1,3 +1,4 @@
 export const routes = {
   signin: './signIn',
+  home: './home',
 };
